Phase 3
Added code to inject data manually once such that it gets reflected in the mongodb. Added features to getUserPosts,getFeedPosts, likePost and verifying with the verifyToken method. Added a postSchema to know who liked,added comment, their details as such using the userId.This ends majority of the backend part. Further will be concentrating on the frontend part.