Phase 5
Added theme.js for the themes that will be used. Made a major mistake by declaring react components by starting with small letters. Remember to declare react components always starting with capital letter. Added the flexBetween.js such that the CSS is enhanced using Material UI and do the changes as required. Completely implemented the NavBar along with the various menu options present in that. In that it is made responsive for both the computer screens and mobile screens.